Make plz clean --nobackground more robust
